restructure project to use submodules
allows for future work not using tui but maybe gui :)
