+++ src/core/timer_runner.rs
@@ -1,5 +1,5 @@
-use crate::timer_commander::TimerCommand;
-use crate::timer_commander::TimerCommand::Stop;
+use crate::core::timer_commander::TimerCommand;
+use crate::core::timer_commander::TimerCommand::Stop;
 use std::sync::mpsc::{Receiver, Sender};
 use std::thread;
 use std::time::{Duration, Instant};
