+++ src/lib.rs
@@ -1,9 +1,3 @@
-pub mod commands;
-pub mod pomodoro_timer;
-mod timer_commander;
-mod timer_database;
-pub mod timer_runner;
-pub mod tui_app;
-
-pub mod models;
-pub mod schema;
+pub mod app;
+pub mod core;
+pub mod db;
